+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Fixed Sky Deck entrance being open despite not having an accessible level.
+- Removed wall that prevented Tails from entering Emerald Coast.
 
 ## [v0.8.1] - 2024-08-29